+++ src/App/src/MyImguiBackends/MyImgui_common.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <functional>
+
+#include "imgui.h"
+
+template <class InputHandler, class Render>
+class MyImgui {
+    MyImgui() {}   
+};
+
+class Imgui_Interface {
+public:
+    Imgui_Interface() {};
+    virtual ~Imgui_Interface() = default;
+
+    virtual void Run(std::function<void()> to_Do) = 0;
+    virtual ImGuiIO& GetIO() = 0;
+    virtual void SetClearColor(float r, float g, float b, float a) = 0;
+    virtual ImVec4& GetClearColor() = 0;
+};
+
